(PC-25511)[PRO] fix: multiple api call in reimbursement page
