+++ pro/src/pages/Reimbursements/ReimbursementBanners.module.scss
@@ -1,0 +1,9 @@
+.banners {
+    display: flex;
+    flex-direction: column;
+    gap: 1rem;
+  
+    &:empty {
+      display: none;
+    }
+  }
